@@ -49,41 +48,31 @@
     }
   },
   "dependencies": {
-    "@graphql-codegen/core": "1.13.0",
-    "@graphql-codegen/typescript": "1.13.0",
-    "apollo-link": "1.2.13",
-    "apollo-link-ws": "1.0.19",
+    "@graphql-codegen/core": "1.13.5",
+    "@graphql-codegen/typescript": "1.13.5",
+    "apollo-link": "1.2.14",
+    "apollo-link-ws": "1.0.20",
     "brace": "0.11.1",
-    "deep-equal": "1.1.1",
     "graphiql": "1.0.0-alpha.0",
     "graphiql-explorer": "0.6.2",
     "graphql": "14.5.8",
     "graphql-voyager": "1.0.0-rc.29",
     "history": "3.3.0",
-    "hoist-non-react-statics": "1.2.0",
     "inflection": "1.12.0",
-    "invariant": "2.2.4",
     "isomorphic-fetch": "2.2.1",
     "jsonwebtoken": "8.5.1",
     "less": "3.11.1",
-    "lru-memoize": "1.0.0",
-    "map-props": "1.0.0",
-    "match-sorter": "2.3.0",
-    "multireducer": "1.0.2",
     "piping": "0.3.2",
-    "pretty-error": "1.2.0",
     "prop-types": "15.7.2",
-    "query-string": "6.11.1",
-    "react": "16.8.6",
+    "react": "16.13.1",
     "react-ace": "8.0.0",
-    "react-autosuggest": "9.4.3",
+    "react-autosuggest": "10.0.2",
     "react-bootstrap": "0.32.4",
     "react-copy-to-clipboard": "5.0.2",
-    "react-dom": "16.8.6",
+    "react-dom": "16.13.1",
     "react-helmet": "5.2.1",
     "react-icons": "3.9.0",
     "react-modal": "3.11.2",
-    "react-notification-system": "0.2.17",
     "react-notification-system-redux": "1.2.0",
     "react-progress-bar-plus": "1.3.1",
     "react-redux": "5.1.2",
