@@ -47,7 +47,7 @@ export default ReactTable => {
       };
     }
 
-    componentWillReceiveProps(newProps) {
+    UNSAFE_componentWillReceiveProps(newProps) {
       if (this.state.resized !== newProps.resized) {
         this.setState({ resized: newProps.resized });
       }
