@@ -49,7 +49,7 @@ class Edit extends React.Component {
     ]);
   }
 
-  componentWillReceiveProps(nextProps) {
+  UNSAFE_componentWillReceiveProps(nextProps) {
     if (
       nextProps.params.remoteSchemaName !== this.props.params.remoteSchemaName
     ) {
