@@ -3,8 +3,9 @@ import styles from '../Common.scss';
 
 const RemoveIcon = ({ className, ...props }) => (
   <i
-    className={`${styles.fontAwosomeClose} fa-lg fa fa-times ${className ||
-      ''}`}
+    className={`${styles.fontAwosomeClose} fa-lg fa fa-times ${
+      className || ''
+    }`}
     {...props}
   />
 );
