@@ -320,37 +320,19 @@ export const getFileExtensionFromFilename = filename => {
 export const getCurrTimeForFileName = () => {
   const currTime = new Date();
 
-  const year = currTime
-    .getFullYear()
-    .toString()
-    .padStart(4, '0');
+  const year = currTime.getFullYear().toString().padStart(4, '0');
 
   const month = (currTime.getMonth() + 1).toString().padStart(2, '0');
 
-  const day = currTime
-    .getDate()
-    .toString()
-    .padStart(2, '0');
+  const day = currTime.getDate().toString().padStart(2, '0');
 
-  const hours = currTime
-    .getHours()
-    .toString()
-    .padStart(2, '0');
+  const hours = currTime.getHours().toString().padStart(2, '0');
 
-  const minutes = currTime
-    .getMinutes()
-    .toString()
-    .padStart(2, '0');
+  const minutes = currTime.getMinutes().toString().padStart(2, '0');
 
-  const seconds = currTime
-    .getSeconds()
-    .toString()
-    .padStart(2, '0');
+  const seconds = currTime.getSeconds().toString().padStart(2, '0');
 
-  const milliSeconds = currTime
-    .getMilliseconds()
-    .toString()
-    .padStart(3, '0');
+  const milliSeconds = currTime.getMilliseconds().toString().padStart(3, '0');
 
   return [year, month, day, hours, minutes, seconds, milliSeconds].join('_');
 };
